# Remove duplicated code from Iterator::{ne, lt, le, gt, ge}