@hellow554 2019-03-18
You could also redirect `parital_cmp` to `cmp`, e.g. and wrapping it into `Some`, because they use exactly the same code :/