@timvermeulen 2019-03-18
@hellow554 Unfortunately we can't forward `partial_cmp` to `cmp` since `cmp` requires that `Self::Item: Ord`. I suppose we could instead forward `cmp` to `partial_cmp` (+ `.unwrap()`), but that doesn't seem worth it. The same applies to `eq`, it can't be forwarded to `cmp`/`partial_cmp`.

There's indeed still some code duplication left but I haven't found a nice way to get rid of it without sacrificing readability. I'm open to suggestions though!