@hellow554 2019-03-19
> I suppose we could instead forward cmp to partial_cmp (+ .unwrap()), but that doesn't seem worth it. 

I think it's worth it, because they share the exact same code (except the `Option` stuff). Code duplication is almost always bad, so why not remove it?