@scottmcm 2019-04-02
Sure, making the way this is implemented match how le/gt/etc are implemented on PartialOrd sounds good to me.

@bors r+ rollup